server/requirements.txt
nso-applet-api
flask
gevent
